persist messages for each room so users can come back to them later
