+++ app/rooms.js
@@ -1,4 +1,21 @@
 const wisdom = require('bob-ross-lipsum')
+const moment = require('moment')
+
+// this should be a database, but for now we'll do
+// persistence in-memory
+var messages = {}
+const rooms = ['John', 'Paul', 'George', 'Ringo']
+rooms.forEach(room => {
+  messages[room] = []
+  for(var i = 0; i < Math.floor(Math.random() * 10) + 1; i ++) {
+    const message  = {
+      content: wisdom(),
+      room: room,
+      time: moment()
+    }
+    messages[room].push(message)
+  }
+})
 
 /**
  * Fetches the messages associated with a room.
@@ -6,13 +23,30 @@ const wisdom = require('bob-ross-lipsum')
  * we can pull from a data store
  */
 function getRoomMessages(req, res) {
-  var messages = []
-  for(var i = 0; i < Math.floor(Math.random() * 10) + 1; i ++) {
-    messages.push(wisdom())
-  }
-  res.send({messages: messages})
+  res.send({messages: messages[req.params.roomId]})
+}
+
+function initializeSocket(io) {
+  io.on('connection', (socket) => {
+    // at the point, users are connected
+    console.log(socket.id, 'connected')
+    socket.on('disconnect', () => {
+      console.log(socket.id, 'disconnected')
+    })
+
+    socket.on('client:message', (data) => {
+      const message  = {
+        content: data.content,
+        room: data.room,
+        time: moment()
+      }
+      messages[data.room].push(message)
+      io.sockets.emit('server:message')
+    })
+  })
 }
 
 module.exports = {
-  getRoomMessages
+  getRoomMessages,
+  initializeSocket
 }
